move the ssl cert to k8s
Since k8s is managing our load balancers through an ingress controller, our SSL cert also needs to be there as well, as the SSL cert is for the load balancer. Previously it was attempting to attach to the static IP address but that was invalid - it needs to point at a load balancer according to google.